+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 20.04.4
+  * Adding metric override feature
   * Fixed a issue in the network request retry logic
   * Removed requirement for any consent to be set before device ID can be changed.
   * Mitigating issue which would happen if push token is refreshed before SDK is initialized.